refactor(capabilities): moving capabilities to documents package
