+++ go/documents/capabilities/burnable.go
