+++ go/documents/capabilities/lockable.go
