+++ go/documents/capabilities/splittable.go
